Add Shell Function, Tweak CSS, Update Readme
Added a shell function, cpnicepdf, that takes two arguments: first the name of the chordpro file, then the name of the output pdf.

Changed .comments in the css file to automatically uppercase the text, to save doing that manually.

Added note in README that {sov}, {eov}, {soc}, and {eoc} must be used.